Implement included scripts.
Implements the `TemplateBuilder` class.
Included scripts are automatically removed if the referenced script is removed and renamed if the referenced script is renamed.

The `Edit script` window now contains an option to specify other scripts to include (prepend) to the triggered script. If your personality uses "helper" scripts to set preferences or create common functions then you can use this option to prepend the contents of your helper scripts before the triggered script is rendered.